requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.3
attrs==25.4.0
Authlib==1.6.6
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
cachetools==7.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
Flask==3.1.2
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
ipinfo==5.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
Naked==0.1.32
ordered-set==4.1.0
packaging==26.0
pillow==12.0.0
propcache==0.4.1
pycparser==2.23
pycryptodome==3.23.0
PyMuPDF==1.25.3
PyMySQL==1.1.2
pyotp==2.9.0
pypdf==6.6.2
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
PyYAML==6.0.3
qrcode==8.2
reportlab==4.4.9
requests==2.32.5
shellescape==3.8.1
soupsieve==2.8.3
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.4
wrapt==2.1.0
yarl==1.22.0
zxcvbn==4.5.0
flask==3.1.2
flask-login==0.6.3
flask-sqlalchemy==3.1.1
gunicorn==25.1.0
jinja2==3.1.6
markupsafe==3.0.3
psycopg2-binary==2.9.11
sqlalchemy==2.0.45
typing-extensions==4.15.0
werkzeug==3.1.4
